Move autoshmap to common spmd_ops
