+++ jaxdecomp/_src/spmd_ops.py
@@ -1,5 +1,6 @@
 from abc import ABCMeta, abstractmethod
 from functools import partial
+from typing import Any, Callable, Hashable
 
 from jax import core
 from jax._src import dispatch
@@ -8,7 +9,61 @@ from jax._src.interpreters import batching
 from jax.experimental.custom_partitioning import custom_partitioning
 from jax.interpreters import mlir, xla
 
+Specs = Any
+AxisName = Hashable
+
+from functools import partial
+
+from jax.experimental.shard_map import shard_map
+from jax.sharding import PartitionSpec as P
+
+from jaxdecomp._src import _jaxdecomp
+
+
+def get_pencil_type():
+  mesh = mesh_lib.thread_resources.env.physical_mesh
+  if mesh.empty:
+    pdims = None
+  else:
+    pdims = mesh.devices.shape[::-1]
+
+  if pdims == (1, 1) or pdims == None:
+    return _jaxdecomp.NO_DECOMP
+  elif pdims[0] == 1:
+    return _jaxdecomp.SLAB_XY
+  elif pdims[1] == 1:
+    return _jaxdecomp.SLAB_YZ
+  else:
+    return _jaxdecomp.PENCILS
+
+
 # Inspired by https://github.com/NVIDIA/TransformerEngine/blob/main/transformer_engine/jax/cpp_extensions.py
+def autoshmap(f: Callable,
+              in_specs: Specs,
+              out_specs: Specs,
+              check_rep: bool = True,
+              auto: frozenset[AxisName] = frozenset(),
+              in_fourrier_space=False) -> Callable:
+  """Helper function to wrap the provided function in a shard map if
+    the code is being executed in a mesh context."""
+  mesh = mesh_lib.thread_resources.env.physical_mesh
+  if mesh.empty:
+    return f
+  else:
+    if in_fourrier_space and 1 in mesh.devices.shape:
+      in_specs, out_specs = switch_specs((in_specs, out_specs))
+    return shard_map(f, mesh, in_specs, out_specs, check_rep, auto)
+
+
+def switch_specs(specs):
+  if isinstance(specs, P):
+    new_axes = tuple(
+        'y' if ax == 'z' else 'z' if ax == 'y' else ax for ax in specs)
+    return P(*new_axes)
+  elif isinstance(specs, tuple):
+    return tuple(switch_specs(sub_spec) for sub_spec in specs)
+  else:
+    raise TypeError("Element must be either a PartitionSpec or a tuple")
 
 
 class BasePrimitive(metaclass=ABCMeta):
